# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "LegalEase REST API",
    "version": "1.0.0",
    "description": "Backend REST API for LegalEase Contract Analysis Platform"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": [
          "name",
          "email",
          "password"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "jane.doe@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "securePassword123"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "jane.doe@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "securePassword123"
          }
        }
      },
      "RefreshRequest": {
        "type": "object",
        "required": [
          "refreshToken"
        ],
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6..."
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11"
          },
          "name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "jane.doe@example.com"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-07-20T18:00:00.000Z"
          }
        }
      },
      "AuthSuccessResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Login successful"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "tokens": {
            "type": "object",
            "properties": {
              "accessToken": {
                "type": "string",
                "example": "eyJhbGciOiJIUzI1..."
              },
              "refreshToken": {
                "type": "string",
                "example": "eyJhbGciOiJIUzI1..."
              }
            }
          }
        }
      },
      "RefreshSuccessResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Token refreshed successfully"
          },
          "tokens": {
            "type": "object",
            "properties": {
              "accessToken": {
                "type": "string",
                "example": "eyJhbGciOiJIUzI1..."
              },
              "refreshToken": {
                "type": "string",
                "example": "eyJhbGciOiJIUzI1..."
              }
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error_code",
          "message"
        ],
        "properties": {
          "error_code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "example": [
              {
                "field": "email",
                "message": "Invalid email format"
              }
            ]
          }
        }
      },
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "ok"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-07-20T18:00:00.000Z"
          }
        }
      }
    }
  },
  "paths": {
    "/api/v1/documents": {
      "post": {
        "summary": "Upload a new contract document",
        "description": "Accepts a PDF or image file, uploads it to S3-compatible storage, creates a database record, and queues analysis.",
        "tags": [
          "Documents"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Contract file to analyze (PDF, PNG, JPEG, JPG, max 10MB)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Document uploaded and analysis job queued successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Document uploaded successfully and analysis has been scheduled."
                    },
                    "document": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                        },
                        "filename": {
                          "type": "string",
                          "example": "employment_contract.pdf"
                        },
                        "s3_key": {
                          "type": "string",
                          "example": "uploads/user-uuid/f47ac10b-58cc-4372-a567-0e02b2c3d479-employment_contract.pdf"
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "uploaded",
                            "processing",
                            "done",
                            "failed"
                          ],
                          "example": "uploaded"
                        },
                        "created_at": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2026-08-03T12:00:00.000Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (no file, unsupported file extension, file too large)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error_code": "BAD_REQUEST",
                  "message": "Invalid file type. Only PDF and images (PNG, JPEG) are allowed."
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error_code": "UNAUTHORIZED",
                  "message": "Authentication token is missing or invalid"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Retrieve a paginated list of documents",
        "description": "Returns a list of uploaded documents belonging to the authenticated user.",
        "tags": [
          "Documents"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of documents to return per page"
          }
        ],
        "responses": {
          "200": {
            "description": "List of documents retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "documents": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid",
                            "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                          },
                          "filename": {
                            "type": "string",
                            "example": "employment_contract.pdf"
                          },
                          "s3_key": {
                            "type": "string",
                            "example": "uploads/user-uuid/f47ac10b-58cc-4372-a567-0e02b2c3d479-employment_contract.pdf"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "uploaded",
                              "processing",
                              "done",
                              "failed"
                            ],
                            "example": "done"
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2026-08-03T12:00:00.000Z"
                          }
                        }
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer",
                          "example": 25
                        },
                        "page": {
                          "type": "integer",
                          "example": 1
                        },
                        "limit": {
                          "type": "integer",
                          "example": 10
                        },
                        "totalPages": {
                          "type": "integer",
                          "example": 3
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error_code": "UNAUTHORIZED",
                  "message": "Authentication token is missing or invalid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{id}": {
      "get": {
        "summary": "Retrieve status of a specific document (Poll endpoint)",
        "description": "Returns the analysis status and metadata of a single document by UUID.",
        "tags": [
          "Documents"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Unique identifier of the document"
          }
        ],
        "responses": {
          "200": {
            "description": "Document status retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "document": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                        },
                        "filename": {
                          "type": "string",
                          "example": "employment_contract.pdf"
                        },
                        "s3_key": {
                          "type": "string",
                          "example": "uploads/user-uuid/f47ac10b-58cc-4372-a567-0e02b2c3d479-employment_contract.pdf"
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "uploaded",
                            "processing",
                            "done",
                            "failed"
                          ],
                          "description": "The analysis status (shows transition stages)",
                          "example": "processing"
                        },
                        "created_at": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2026-08-03T12:00:00.000Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed"
          },
          "404": {
            "description": "Document not found or user lacks permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error_code": "NOT_FOUND",
                  "message": "Document not found or access denied"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "summary": "Register a new user",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              },
              "example": {
                "name": "Jane Doe",
                "email": "jane.doe@example.com",
                "password": "securePassword123"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSuccessResponse"
                },
                "example": {
                  "message": "User registered successfully",
                  "user": {
                    "id": "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11",
                    "name": "Jane Doe",
                    "email": "jane.doe@example.com",
                    "created_at": "2026-07-20T18:00:00.000Z"
                  },
                  "tokens": {
                    "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6...",
                    "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6..."
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or missing fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error_code": "VALIDATION_ERROR",
                  "message": "Validation failed",
                  "details": [
                    {
                      "field": "email",
                      "message": "Invalid email address format"
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "User already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error_code": "CONFLICT",
                  "message": "A user with this email address already exists"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "summary": "Authenticate user and issue tokens",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "email": "jane.doe@example.com",
                "password": "securePassword123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSuccessResponse"
                },
                "example": {
                  "message": "Login successful",
                  "user": {
                    "id": "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11",
                    "name": "Jane Doe",
                    "email": "jane.doe@example.com",
                    "created_at": "2026-07-20T18:00:00.000Z"
                  },
                  "tokens": {
                    "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6...",
                    "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6..."
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error_code": "VALIDATION_ERROR",
                  "message": "Validation failed"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error_code": "UNAUTHORIZED",
                  "message": "Invalid email or password"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "summary": "Refresh access token using stateful refresh token",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              },
              "example": {
                "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6..."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshSuccessResponse"
                },
                "example": {
                  "message": "Token refreshed successfully",
                  "tokens": {
                    "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6...",
                    "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6..."
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error_code": "VALIDATION_ERROR",
                  "message": "Validation failed"
                }
              }
            }
          },
          "401": {
            "description": "Refresh token invalid, expired, or revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error_code": "UNAUTHORIZED",
                  "message": "Refresh token is invalid, expired, or has been revoked"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Healthcheck endpoint for containers and monitoring",
        "tags": [
          "System"
        ],
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": []
}